seed_data/requirements.txt
# Requirements for c15-court-transcript-pipeline
pytest
pylint
pytest-mock
pytest-cov
requests
pandas
python-dotenv
pyarrow
boto3
moto
psycopg2
openai
beautifulsoup4
lxml
aiohttp